+++ Week3/Lunch-Recommendation/src/InitialPage.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
-import InitPage from './components/InitPage';
-import RecommendationPage from './components/RecommendationPage';
-import './App.css';
+import SelectRecommendType from './components/SelectRecommendType';
+import './style.css';
+import RenderToSelect from './components/RenderToSelect';
 
-const App = () => {
+const InitialPage = () => {
   const [recommendationStart, setRecommendationStart] = useState(false);
   const [selectedOption, setSelectedOption] = useState(null);
   const handleReset = () => {
@@ -16,10 +16,10 @@ const App = () => {
       <h1>😋 점메추 😋</h1>
       <button onClick={() => handleReset()} type='button'>처음으로</button>
       {recommendationStart ? (
-        <RecommendationPage selectedOption={selectedOption} setRecommendationStart={setRecommendationStart} />
+        <RenderToSelect selectedOption={selectedOption} setRecommendationStart={setRecommendationStart} />
       ) : (
         <>
-          <InitPage
+          <SelectRecommendType
             recommendationStart={recommendationStart}
             setRecommendationStart={setRecommendationStart}
             selectedOption={selectedOption}
@@ -31,4 +31,4 @@ const App = () => {
   );
 }
 
-export default App;
+export default InitialPage;
